sprint demo 4

+++ Graphics/include/Animation/Animation.cpp
@@ -13,7 +13,7 @@ std::pair<KeyFrame const&, KeyFrame const&> Animation::getKeyFrames(float durati
     int prevKeyFrame = 0;
     int nextKeyFrame = 0;
 
-    for (int i = 0, size = keyFrames.size(); i < size; i++)
+    for (int i = 0, size = (int)keyFrames.size(); i < size; i++)
     {
         if (keyFrames[i].getTimeStamp() < duration)
         {
